# Optimize builder and HTTP policy allocations