Before:
```
 Method |     Mean |    Error |   StdDev |  Op/s |    Gen 0 |    Gen 1 |    Gen 2 | Allocated |
------- |---------:|---------:|---------:|------:|---------:|---------:|---------:|----------:|
    Dfa | 324.0 ms | 6.375 ms | 14.52 ms | 3.087 | 562.5000 | 500.0000 | 375.0000 | 130.02 MB |
```

After:
```
 Method |     Mean |    Error |   StdDev |  Op/s |    Gen 0 |    Gen 1 |    Gen 2 | Allocated |
------- |---------:|---------:|---------:|------:|---------:|---------:|---------:|----------:|
    Dfa | 279.5 ms | 5.582 ms | 7.452 ms | 3.578 | 687.5000 | 500.0000 | 312.5000 | 115.25 MB |
```