@JamesNK 2018-09-05
Removing jump table builders and their array knocked another 5 megs off allocations 🍰